+++ CHANGELOG.md
@@ -2,4 +2,5 @@ CHANGELOG for Sulu Pricing Bundle
 =================================
 
 * dev-develop
+    * FEATURE Added handling of recurring-prices
     * FEATURE Added handling of gross-prices